requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.22.0

httpx[http2]>=0.24.1
redis>=4.5.5
python-dotenv>=1.0.0

google-genai>=0.2.0,<0.3.0

pydantic>=2.0,<3.0

aio-pika>=9.4.0
